refactor: User 데이터 Enum 폴더 분리
