+++ src/main/java/com/oreumi/pet_trip_service/model/Enum/AuthProvider.java
@@ -1,4 +1,4 @@
-package com.oreumi.pet_trip_service.model;
+package com.oreumi.pet_trip_service.model.Enum;
 
 public enum AuthProvider {
     LOCAL,
