+++ src/main/java/com/oreumi/pet_trip_service/model/User.java
@@ -1,5 +1,7 @@
 package com.oreumi.pet_trip_service.model;
 
+import com.oreumi.pet_trip_service.model.Enum.AuthProvider;
+import com.oreumi.pet_trip_service.model.Enum.UserStatus;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
